Show the watch panel as a sheet
It was a popover anchored to its trigger, which meant two placements, a
click-outside listener and an Escape listener — all so the same panel could hang
off the header on a desktop and off the tab bar on a phone. As a sheet it is one
shape in both places, and <dialog> already knows how to do the rest.
