+++ src/lib/shared/components/app-header/app-header.svelte
@@ -2,7 +2,7 @@
   import { page } from "$app/state";
   import { Icon } from "$lib/shared/components/icon";
   import { authModel } from "$lib/modules/auth/model";
-  import WatchPopover from "$lib/modules/device/components/watch-popover.svelte";
+  import WatchSheet from "$lib/modules/device/components/watch-sheet.svelte";
   import { Avatar } from "$lib/shared/components/avatar";
   import { Menu, MenuItem } from "$lib/shared/components/menu";
 
@@ -22,14 +22,14 @@
         {item.title}
       </a>
     {/each}
-    <WatchPopover>
+    <WatchSheet>
       {#snippet trigger({ open, toggle, connected })}
         <button class="watch-btn" class:active={open} onclick={toggle} aria-label="Watch">
           <Icon name="watch" size={18} />
           {#if connected}<span class="dot"></span>{/if}
         </button>
       {/snippet}
-    </WatchPopover>
+    </WatchSheet>
   </nav>
   <div class="user">
     <a
